- Implementación jacoco y sonar - Se finalizan pruebas unitarias - Se completo la documentación de las clases - Validación cobertura

+++ src/main/java/com/nisum/pruebanisum/utilities/JwtTokenUtilities.java
@@ -14,7 +14,7 @@ import java.util.function.Function;
 @Component
 public class JwtTokenUtilities {
 
-    public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+    public static final long JWT_TOKEN_VALIDITY = (long) 5 * 60 * 60;
     @Value("${jwt.secret}")
     private String secret;
